+++ packages/core/src/decorators/ManyToOne.ts
@@ -55,4 +55,10 @@ export interface ManyToOneOptions<Owner, Target> extends ReferenceOptions<Owner,
 
   /** Set the constraint type. Immediate constraints are checked for each statement, while deferred ones are only checked at the end of the transaction. Only for postgres unique constraints. */
   deferMode?: DeferMode | `${DeferMode}`;
+
+  /** Enable/disable foreign key constraint creation on this relation */
+  createForeignKeyConstraint?: boolean;
+
+  /** Set a custom foreign key constraint name, overriding NamingStrategy.indexName(). */
+  foreignKeyName?: string;
 }