+++ packages/core/src/entity/Collection.ts
@@ -15,10 +15,10 @@ import type {
 import { ArrayCollection } from './ArrayCollection';
 import { DataloaderUtils, Utils } from '../utils';
 import { ValidationError } from '../errors';
-import { type QueryOrderMap, ReferenceKind, DataloaderType } from '../enums';
+import { DataloaderType, type QueryOrderMap, ReferenceKind } from '../enums';
 import { Reference } from './Reference';
 import type { Transaction } from '../connections/Connection';
-import type { FindOptions, CountOptions, LoadHint } from '../drivers/IDatabaseDriver';
+import type { CountOptions, FindOptions, LoadHint } from '../drivers/IDatabaseDriver';
 import { helper } from './wrap';
 import type { EntityLoaderOptions } from './EntityLoader';
 
@@ -301,14 +301,22 @@ export class Collection<T extends object, O extends object = object> extends Arr
       const order = [...this.items]; // copy order of references
       const orderBy = this.createOrderBy(options.orderBy);
       const customOrder = orderBy.length > 0;
-      // eslint-disable-next-line dot-notation
-      const items: TT[] = await em['colLoader'].load([
+      const pivotTable = this.property.kind === ReferenceKind.MANY_TO_MANY && em.getPlatform().usesPivotTable();
+      const loader = pivotTable ? 'colLoaderMtoN' : 'colLoader';
+      const items: TT[] = await em[loader].load([
         this,
         { ...options, orderBy },
       ]);
 
-      if (!customOrder) {
-        this.reorderItems(items, order);
+      if (this.property.kind === ReferenceKind.MANY_TO_MANY) {
+        this.initialized = true;
+        this.dirty = false;
+
+        if (!customOrder) {
+          this.reorderItems(items, order);
+        }
+
+        return this as unknown as LoadedCollection<Loaded<TT, P>>;
       }
 
       this.items.clear();