+++ packages/core/src/metadata/MetadataDiscovery.ts
@@ -352,7 +352,8 @@ export class MetadataDiscovery {
         MetadataStorage.getMetadata(entity.meta.className, filepath);
       }
 
-      return entity;
+      const meta = Utils.copy(entity.meta, false);
+      return EntitySchema.fromMetadata(meta);
     }
 
     const path = entity[MetadataStorage.PATH_SYMBOL];
@@ -825,6 +826,7 @@ export class MetadataDiscovery {
       autoincrement: false,
       updateRule: prop.updateRule,
       deleteRule: prop.deleteRule,
+      createForeignKeyConstraint: prop.createForeignKeyConstraint,
     } as EntityProperty;
 
     if (selfReferencing && !this.platform.supportsMultipleCascadePaths()) {
@@ -1241,7 +1243,7 @@ export class MetadataDiscovery {
     }, {} as Dictionary);
   }
 
-  private getDefaultVersionValue(prop: EntityProperty): string {
+  private getDefaultVersionValue(meta: EntityMetadata, prop: EntityProperty): string {
     if (typeof prop.defaultRaw !== 'undefined') {
       return prop.defaultRaw;
     }
@@ -1251,7 +1253,10 @@ export class MetadataDiscovery {
       return '' + this.platform.quoteVersionValue(prop.default as number, prop);
     }
 
-    if (prop.type.toLowerCase() === 'date') {
+    this.initCustomType(meta, prop, true);
+    const type = prop.customType?.runtimeType ?? prop.runtimeType ?? prop.type;
+
+    if (type === 'Date') {
       prop.length ??= this.platform.getDefaultVersionLength();
       return this.platform.getCurrentTimestampSQL(prop.length);
     }
@@ -1303,7 +1308,7 @@ export class MetadataDiscovery {
       return;
     }
 
-    if (prop.customType instanceof ArrayType && Array.isArray(prop.default)) {
+    if (Array.isArray(prop.default) && prop.customType) {
       val = prop.customType.convertToDatabaseValue(prop.default, this.platform)!;
     }
 
@@ -1330,7 +1335,7 @@ export class MetadataDiscovery {
     if (prop.version) {
       this.initDefaultValue(prop);
       meta.versionProperty = prop.name;
-      prop.defaultRaw = this.getDefaultVersionValue(prop);
+      prop.defaultRaw = this.getDefaultVersionValue(meta, prop);
     }
 
     if (prop.concurrencyCheck && !prop.primary) {
@@ -1338,7 +1343,7 @@ export class MetadataDiscovery {
     }
   }
 
-  private initCustomType(meta: EntityMetadata, prop: EntityProperty): void {
+  private initCustomType(meta: EntityMetadata, prop: EntityProperty, simple = false): void {
     // `prop.type` might be actually instance of custom type class
     if (Type.isMappedType(prop.type) && !prop.customType) {
       prop.customType = prop.type;
@@ -1351,11 +1356,19 @@ export class MetadataDiscovery {
       prop.type = prop.customType.constructor.name;
     }
 
+    if (simple) {
+      return;
+    }
+
     if (!prop.customType && ['json', 'jsonb'].includes(prop.type?.toLowerCase())) {
       prop.customType = new JsonType();
     }
 
     if (prop.kind === ReferenceKind.SCALAR && !prop.customType && prop.columnTypes && ['json', 'jsonb'].includes(prop.columnTypes[0])) {
+      prop.customType = new JsonType();
+    }
+
+    if (prop.kind === ReferenceKind.EMBEDDED && !prop.customType && (prop.object || prop.array)) {
       prop.customType = new JsonType();
     }
 
@@ -1397,6 +1410,8 @@ export class MetadataDiscovery {
       } else {
         prop.runtimeType ??= prop.customType.runtimeType as typeof prop.runtimeType;
       }
+    } else if (prop.runtimeType === 'object') {
+      prop.runtimeType = mappedType.runtimeType as typeof prop.runtimeType;
     } else {
       prop.runtimeType ??= mappedType.runtimeType as typeof prop.runtimeType;
     }