+++ tests/entities-schema/FooBar4.ts
@@ -1,4 +1,4 @@
-import type { OptionalProps } from '@mikro-orm/core';
+import { IntegerType, OptionalProps } from '@mikro-orm/core';
 import { ArrayType, BlobType, EntitySchema, JsonType, Uint8ArrayType } from '@mikro-orm/core';
 import type { IFooBaz4, IBaseEntity5 } from './index';
 import { BaseEntity5 } from './index';
@@ -23,7 +23,7 @@ export const FooBar4 = new EntitySchema<IFooBar4, IBaseEntity5>({
     name: { type: 'string', default: 'asd' },
     baz: { kind: '1:1', entity: 'FooBaz4', orphanRemoval: true, nullable: true },
     fooBar: { kind: '1:1', entity: 'FooBar4', nullable: true },
-    version: { type: 'number', version: true },
+    version: { type: IntegerType, version: true },
     blob: { type: BlobType, nullable: true },
     blob2: { type: Uint8ArrayType, nullable: true },
     array: { type: new ArrayType(i => +i), nullable: true },